+++ HISTORY.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2019-10-30
+### Added
+- Added `generate_default_bootstraptheme` management command
+
+### Changed
+- Added more informative error messages when using template tags without a theme
+
 ## [0.2.0] - 2019-10-02
 ### Changed
 - Updated to Bootstrap 4.3.1